oddo_hack/oddo/rewear/server/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
mysqlclient==2.2.1
python-dotenv==1.0.1
Pillow==10.2.0
marshmallow==3.20.2
argon2-cffi==23.1.0
pytz==2024.1
gunicorn==21.2.0
python-socketio==5.11.0
eventlet==0.35.2
flask-socketio==5.3.6
python-socketio==5.11.0